refactor(Auth) Lazy load Auth scene
